Created separate files for lexer and parser
